requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==3.5.0
bleach==5.0.1
blinker==1.7.0
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==2.1.0
chart-studio==1.1.0
ckzg==2.1.1
click==8.1.3
colorama==0.4.5
colorlover==0.3.0
ctransformers==0.2.27
cycler==0.11.0
cytoolz==1.0.1
dataclasses-json==0.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==5.0.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-suit==0.2.28
djangorestframework==3.15.2
entrypoints==0.4
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.16.1
filelock==3.12.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
fonttools==4.37.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google==3.0.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==2.0.2
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
hexbytes==1.3.1
huggingface-hub==0.20.1
idna==2.8
importlib-metadata==6.11.0
ipfshttpclient==0.7.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.15.0
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
langchain==0.0.352
langchain-community==0.0.5
langchain-core==0.1.2
langsmith==0.0.72
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.5.3
mdurl==0.1.2
mistune==2.0.4
mpmath==1.3.0
msgpack==1.1.0
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.5
netaddr==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.23.1
opacus==1.5.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt_einsum==3.4.0
packaging==23.2
pandas==1.4.4
pandocfilters==1.5.0
parsimonious==0.10.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.1
plotly==5.14.1
portpicker==1.2.0
promise==2.3
prompt-toolkit==1.0.18
proto-plus==1.23.0
protobuf==4.25.1
psycopg2==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.23.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.13.0
PyOpenGL @ file:///C:/Users/Lenovo/Downloads/PyOpenGL-3.1.6-cp310-cp310-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/Lenovo/Downloads/PyOpenGL_accelerate-3.1.6-cp310-cp310-win_amd64.whl
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.18.1
python-box==7.1.1
python-dateutil==2.8.2
pytz==2024.1
pyunormalize==16.0.0
pywin32==304
PyYAML==6.0.1
pyzmq==23.2.1
redis==5.1.0
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rich==13.7.0
rlp==4.1.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.3
simplegeneric==0.8.1
six==1.12.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.7
sqlparse==0.4.4
streamlit==1.29.0
stringcase==1.2.0
sympy==1.12
tbb==2021.13.0
tenacity==8.2.2
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
torchaudio==2.1.2+cpu
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.3.0
transformers==4.36.2
txaio==23.1.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.25.0
validators==0.22.0
varint==1.0.2
vine==5.1.0
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.13
web3==7.12.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
